rename: enum 파일 분리 - loginType - storageType - textFieldType - mainButtonType - selectButtonType
