+++ lib/components/main_button.dart
@@ -1,15 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:nutripic/utils/enums/main_button_type.dart';
 import 'package:nutripic/utils/palette.dart';
-
-/// Main Button 타입
-enum MainButtonType {
-  primary(color: Palette.sub, textColor: Palette.white),
-  secondary(color: Palette.gray100, textColor: Palette.gray400);
-
-  final Color color;
-  final Color textColor;
-  const MainButtonType({required this.color, required this.textColor});
-}
 
 class MainButton extends StatelessWidget {
   final String label;
