+++ lib/utils/enums/login_type.dart
@@ -1,0 +1,2 @@
+/// 로그인 방식 (이메일, 카카오, 구글, 애플)
+enum LoginType { email, kakao, google, apple }
