+++ lib/utils/enums/text_field_type.dart
@@ -1,0 +1,1 @@
+enum TextFieldType { email, password, text }
